+++ src/test/java/org/organicdesign/fp/LazyRefTest.java
@@ -17,6 +17,8 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import org.organicdesign.fp.function.Function0;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 import static org.junit.Assert.assertEquals;
 
 @RunWith(JUnit4.class)
@@ -24,26 +26,26 @@ public class LazyRefTest {
 
     @Test
     public void testLazyRef() {
-        Mutable.IntRef intRef = Mutable.IntRef.of(3);
-        Function0<Integer> f = () -> intRef.increment().value();
-        assertEquals(f.apply(), new Integer(4));
-        assertEquals(f.apply(), new Integer(5));
-        assertEquals(f.apply(), new Integer(6));
+        AtomicInteger intRef = new AtomicInteger(3);
+        Function0<Integer> f = () -> intRef.incrementAndGet();
+        assertEquals(new Integer(4), f.apply());
+        assertEquals(new Integer(5), f.apply());
+        assertEquals(new Integer(6), f.apply());
 
         LazyRef<Integer> lr = LazyRef.of(f);
 
         assertEquals("LazyRef(*not-computed-yet*)", lr.toString());
 
-        assertEquals(lr.get(), new Integer(7));
+        assertEquals(new Integer(7), lr.get());
 
         assertEquals("LazyRef(7)", lr.toString());
 
-        assertEquals(f.apply(), new Integer(8));
+        assertEquals(new Integer(8), f.apply());
         intRef.set(-1);
 
-        assertEquals(lr.get(), new Integer(7));
-        assertEquals(lr.get(), new Integer(7));
-        assertEquals(lr.get(), new Integer(7));
+        assertEquals(new Integer(7), lr.get());
+        assertEquals(new Integer(7), lr.get());
+        assertEquals(new Integer(7), lr.get());
     }
 
     @Test (expected = IllegalArgumentException.class)